New models

+++ bll/BllService/BllBorrowRequestsService.cs
@@ -1,8 +1,12 @@
 ﻿using AutoMapper;
 using BLL.BllModels;
 using BLL.IBll;
-using dal.models;
+
 using DAL;
+using DAL.IDal;
+using DAL.models;
+using System.Linq;
+
 
 using static BLL.Exeptions.BorrowRequestExeptions;
 
@@ -117,7 +121,8 @@ namespace BLL.BllService
 
                 foreach (var borrowApprovalRequest in borrowApprovalRequests)
                 {
-                    itemIds.Add(borrowApprovalRequest.ItemId);
+                      itemIds.Add((int)borrowApprovalRequest.ItemId);
+                 
 
                 }
 
@@ -127,9 +132,6 @@ namespace BLL.BllService
 
                 }
 
-                // קריאת רשימת הפריטים לפי itemIds
-                /*                List<Item> items = await _dalManager.Items.Read(i => itemIds.Contains(i.Id));
-                */
                 List<Item> items = await _dalManager.items.Read(i => itemIds.Contains(i.Id));
                 return mapper.Map<List<Item>, List<BllItem>>(items);
             }
